Introduce GetFilesystems and Getusers
To get the full list of filesystems and
users on the guest machine, this commit
introduces the two endpoint required in
the virt-launcher.

+++ pkg/virt-launcher/virtwrap/api/schema.go
@@ -150,7 +150,7 @@ type Filesystem struct {
 type User struct {
 	Name      string
 	Domain    string
-	LoginTime int
+	LoginTime float64
 }
 
 // DomainGuestInfo represent guest agent info for specific domain
